fix few programs

+++ Homework/Marianna/Loops/CalculateSumByFormula/Program.cs
@@ -11,13 +11,13 @@ namespace CalculateSumByFormula
             Console.Write("Please enter x: ");
             int x = Int32.Parse(Console.ReadLine());
             double sum = 1;
-            int factorial = 1;
+            int  factorial = 1;
             int currentX = 1;
             for (int i = 1; i <= n; i++)
             {
                 factorial *= i;
                 currentX *= x;
-                sum += factorial / currentX;
+                sum += (double)factorial / currentX;
             }
             Console.WriteLine("Result: " + sum);
         }
